ファイル構成の変更
